Migrate package dependency to dotnet core build tasks, and delete legacy sources
