--- src/Tasks/LockFileToMSBuildSamples/SampleTargets.targets
@@ -1,91 +1,0 @@
-﻿<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <UsingTask TaskName="Microsoft.MSBuild.LockFile.Tasks.LockFileToMSBuild" AssemblyFile="..\Microsoft.MSBuild.LockFile.Tasks\bin\Debug\Microsoft.MSBuild.LockFile.Tasks.dll" />
-
-  <PropertyGroup Condition="'$(ProjectLockFile)' == ''">
-    <ProjectLockFile>$(MSBuildThisFileDirectory)NetFramework4.5.2\project.lock.json</ProjectLockFile>
-  </PropertyGroup>
-
-  <Target Name="RunLockFileToMSBuild">
-    <Message Text="     -> (Raising Lock File: ProjectLockFile=$(ProjectLockFile))" Importance="high"/>
-
-    <LockFileToMSBuild 
-      ProjectLockFile="$(ProjectLockFile)">
-
-      <Output TaskParameter="LockFileLibraries" ItemName="LockFileLibraries" />
-      <Output TaskParameter="LockFileTargets" ItemName="LockFileTargets" />
-      <Output TaskParameter="LockFileTargetLibraries" ItemName="LockFileTargetLibraries" />
-      <Output TaskParameter="LockFileItems" ItemName="LockFileItems" />
-      <Output TaskParameter="PackageDependencies" ItemName="PackageDependencies" />
-    </LockFileToMSBuild>
-
-  </Target>
-
-  <!-- Sample 1: Display Counts of all Items -->
-  <Target Name="DisplayItemCounts" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="LockFileLibraries (@(LockFileLibraries->Count()))" Importance="high" />
-    <Message Text="LockFileTargets (@(LockFileTargets->Count()))" Importance="high" />
-    <Message Text="LockFileTargetLibraries (@(LockFileTargetLibraries->Count()))" Importance="high" />
-    <Message Text="LockFileItems (@(LockFileItems->Count()))" Importance="high" />
-    <Message Text="PackageDependencies (@(PackageDependencies->Count()))" Importance="high" />
-  </Target>
-
-  <!-- Sample 2: Display The RIDs in the LockFile -->
-  <Target Name="DisplayAllRIDs" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="RID (%(LockFileTargets.RuntimeIdentifier))" Importance="high" />
-  </Target>
-
-  <!-- Sample 3: Display The Targets in the LockFile -->
-  <Target Name="DisplayAllTargets" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="%(LockFileTargets.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample 4: Display libraries in a specific target -->
-  <Target Name="DisplayLibrariesPerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMLibraries Include="@(LockFileTargetLibraries->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMLibraries>
-    </ItemGroup>
-    
-    <Message Text="%(TFMLibraries.Identity), %(Version)" Importance="high" />
-  </Target>
-
-  <!-- Sample 5: Display Compile Time Assemblies in a specific Target -->
-  <Target Name="DisplayCompilePerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMFileItems Include="@(LockFileItems->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMFileItems>
-      <TFMCompileFileItems Include="@(TFMFileItems->WithMetadataValue('FileGroup', 'CompileTimeAssembly'))"></TFMCompileFileItems>
-    </ItemGroup>
-
-    <Message Text="%(TFMCompileFileItems.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample 6: Display libraries in a specific target and their dependencies -->
-  <Target Name="DisplayLibrariesPerTargetAndDeps" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMDeps Include="@(PackageDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMDeps>
-    </ItemGroup>
-
-    <Message Text="%(ParentTargetLibrary) -> %(TFMDeps.Identity), %(TFMDeps.Version)" Importance="high" />
-  </Target>
-
-  <!-- Sample Z: Display all Identities created by LockFileToMSBuild task -->
-  <Target Name="DisplayAllIdentities" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="LockFileLibraries (@(LockFileLibraries->Count()))" Importance="high" />
-    <Message Text="  - %(LockFileLibraries.Identity)" Importance="high" />
-
-    <Message Text="LockFileTargets (@(LockFileTargets->Count()))" Importance="high" />
-    <Message Text="  - %(LockFileTargets.Identity)" Importance="high" />
-
-    <Message Text="LockFileTargetLibraries (@(LockFileTargetLibraries->Count()))" Importance="high" />
-    <Message Text="  - %(LockFileTargetLibraries.Identity)" Importance="high" />
-
-    <Message Text="LockFileItems (@(LockFileItems->Count()))" Importance="high" />
-    <Message Text="  - %(LockFileItems.Identity)" Importance="high" />
-
-    <Message Text="PackageDependencies (@(PackageDependencies->Count()))" Importance="high" />
-    <Message Text="  - %(PackageDependencies.Identity)" Importance="high" />
-
-    <!--<Message Text="LockFileItems (@(LockFileItems->Count()))" Importance="high" />
-    <Message Text="  - %(LockFileItems.Identity), %(LockFileItems.FileGroup), %(LockFileItems.TargetLibrary), %(LockFileItems.ParentTarget), " Importance="high" />-->
-  </Target>
-  
-</Project>
