--- src/Tasks/LockFileToMSBuildSamples/SampleTargets2.targets
@@ -1,113 +1,0 @@
-﻿<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <UsingTask TaskName="Microsoft.MSBuild.LockFile.Tasks.LockFileToMSBuild2" AssemblyFile="..\Microsoft.MSBuild.LockFile.Tasks\bin\Debug\Microsoft.MSBuild.LockFile.Tasks.dll" />
-
-  <PropertyGroup Condition="'$(ProjectLockFile)' == ''">
-    <ProjectLockFile>$(MSBuildThisFileDirectory)NetFramework4.5.2\project.lock.json</ProjectLockFile>
-  </PropertyGroup>
-
-  <Target Name="RunLockFileToMSBuild">
-    <Message Text="     -> (Raising Lock File: ProjectLockFile=$(ProjectLockFile))" Importance="high"/>
-
-    <LockFileToMSBuild2 
-      ProjectLockFile="$(ProjectLockFile)">
-
-      <Output TaskParameter="TargetDefinitions" ItemName="TargetDefinitions" />
-      <Output TaskParameter="PackageDefinitions" ItemName="PackageDefinitions" />
-      <Output TaskParameter="FileDefinitions" ItemName="FileDefinitions" />
-      <Output TaskParameter="PackageDependencies" ItemName="PackageDependencies" />
-      <Output TaskParameter="FileDependencies" ItemName="FileDependencies" />
-    </LockFileToMSBuild2>
-
-  </Target>
-
-  <!-- Sample 1: Display Counts of all Items -->
-  <Target Name="DisplayItemCounts" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="TargetDefinitions (@(TargetDefinitions->Count()))" Importance="high" />
-    <Message Text="PackageDefinitions (@(PackageDefinitions->Count()))" Importance="high" />
-    <Message Text="FileDefinitions (@(FileDefinitions->Count()))" Importance="high" />
-    <Message Text="PackageDependencies (@(PackageDependencies->Count()))" Importance="high" />
-    <Message Text="FileDependencies (@(FileDependencies->Count()))" Importance="high" />
-  </Target>
-
-  <!-- Sample 2: Display The RIDs in the LockFile -->
-  <Target Name="DisplayAllRIDs" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="RID (%(TargetDefinitions.RuntimeIdentifier))" Importance="high" />
-  </Target>
-
-  <!-- Sample 3: Display The Targets in the LockFile -->
-  <Target Name="DisplayAllTargets" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="%(TargetDefinitions.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample 4: Display Package Dependencies in a specific target -->
-  <Target Name="DisplayPackageDependenciesPerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMLibraries Include="@(PackageDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMLibraries>
-    </ItemGroup>
-    
-    <Message Text="%(TFMLibraries.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample 4b: Display definitions for the Package Dependencies in a specific target -->
-  <Target Name="DisplayPackageDefinitionsPerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMLibraries Include="@(PackageDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMLibraries>
-      <TFMDefinitionsEx Include="@(PackageDefinitions)" Exclude="@(TFMLibraries)"></TFMDefinitionsEx>
-      <TFMDefinitions Include="@(PackageDefinitions)" Exclude="@(TFMDefinitionsEx)"></TFMDefinitions>
-    </ItemGroup>
-
-    <Message Text="%(TFMDefinitions.Identity) [%(Path)] [%(Type)]" Importance="high" />
-  </Target>
-
-  <!-- Sample 5: Display Compile Time Assemblies in a specific Target -->
-  <Target Name="DisplayCompilePerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMFileItems Include="@(FileDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMFileItems>
-      <TFMCompileFileItems Include="@(TFMFileItems->WithMetadataValue('FileGroup', 'CompileTimeAssembly'))"></TFMCompileFileItems>
-    </ItemGroup>
-
-    <Message Text="TFMCompileFileItems (@(TFMCompileFileItems->Count()))" Importance="high" />
-    <Message Text="%(TFMCompileFileItems.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample 5b: Display Path metadata for Compile Time Assemblies in a specific Target -->
-  <Target Name="DisplayCompileDefnPerTarget" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMFileItems Include="@(FileDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMFileItems>
-      <TFMCompileFileItems Include="@(TFMFileItems->WithMetadataValue('FileGroup', 'CompileTimeAssembly'))"></TFMCompileFileItems>
-      <TFMCompileFileDefinitionsEx Include="@(FileDefinitions)" Exclude="@(TFMCompileFileItems)"></TFMCompileFileDefinitionsEx>
-      <TFMCompileFileDefinitions Include="@(FileDefinitions)" Exclude="@(TFMCompileFileDefinitionsEx)"></TFMCompileFileDefinitions>
-    </ItemGroup>
-
-    <Message Text="%(TFMCompileFileDefinitions.Identity) == %(ResolvedPath)" Importance="high" />
-  </Target>
-
-  <!-- Sample 6: Display libraries in a specific target and their dependencies -->
-  <Target Name="DisplayLibrariesPerTargetAndDeps" Condition="'$(TargetMoniker)' != ''" DependsOnTargets="RunLockFileToMSBuild">
-    <ItemGroup>
-      <TFMDeps Include="@(PackageDependencies->WithMetadataValue('ParentTarget', '$(TargetMoniker)'))"></TFMDeps>
-    </ItemGroup>
-
-    <Message Text="%(ParentPackage) -> %(TFMDeps.Identity)" Importance="high" />
-  </Target>
-
-  <!-- Sample Z: Display all Identities created by LockFileToMSBuild task -->
-  <Target Name="DisplayAllIdentities" DependsOnTargets="RunLockFileToMSBuild">
-    <Message Text="TargetDefinitions (@(TargetDefinitions->Count()))" Importance="high" />
-    <Message Text="  - %(TargetDefinitions.Identity)" Importance="high" />
-
-    <Message Text="PackageDefinitions (@(PackageDefinitions->Count()))" Importance="high" />
-    <Message Text="  - %(PackageDefinitions.Identity)" Importance="high" />
-
-    <Message Text="FileDefinitions (@(FileDefinitions->Count()))" Importance="high" />
-    <Message Text="  - %(FileDefinitions.Identity)" Importance="high" />
-
-    <Message Text="PackageDependencies (@(PackageDependencies->Count()))" Importance="high" />
-    <Message Text="  - %(PackageDependencies.Identity) [%(PackageDependencies.ParentTarget)] [%(PackageDependencies.ParentPackage)]" Importance="high" />
-
-    <Message Text="FileDependencies (@(FileDependencies->Count()))" Importance="high" />
-    <Message Text="  - %(FileDependencies.Identity) [%(FileDependencies.FileGroup)] [%(FileDependencies.ParentTarget)] [%(FileDependencies.ParentPackage)]" Importance="high" />
-
-  </Target>
-  
-</Project>
