+++ src/Tasks/Microsoft.DotNet.Core.Build.Tasks/MetadataKeys.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Microsoft.MSBuild.LockFile.Tasks
+namespace Microsoft.DotNet.Core.Build.Tasks
 {
     internal static class MetadataKeys
     {
