+++ src/Tasks/Microsoft.DotNet.Core.Build.Tasks/Microsoft.DotNet.Core.Build.Tasks.csproj
@@ -17,7 +17,10 @@
   <ItemGroup>
     <None Include="Microsoft.DotNet.Core.Sdk.targets" />
     <None Include="Microsoft.DotNet.Core.Sdk.props" />
+    <None Include="Microsoft.PackageDependencyResolution.targets" />
     <None Include="project.json" />
+    <Compile Include="FileGroup.cs" />
+    <Compile Include="MetadataKeys.cs" />
     <Compile Include="ResolvePackageDependencies.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
