--- src/Tasks/Microsoft.MSBuild.LockFile.Tasks/Microsoft.MSBuild.LockFile.Tasks.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.MSBuild.LockFile.Tasks", "Microsoft.MSBuild.LockFile.Tasks.csproj", "{1C024D50-3703-4EB5-B9AE-DFB9D147E947}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LockFileToMSBuildSamples", "..\LockFileToMSBuildSamples\LockFileToMSBuildSamples.csproj", "{3837A8AF-0D68-4552-9255-058006EE85D4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1C024D50-3703-4EB5-B9AE-DFB9D147E947}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C024D50-3703-4EB5-B9AE-DFB9D147E947}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C024D50-3703-4EB5-B9AE-DFB9D147E947}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C024D50-3703-4EB5-B9AE-DFB9D147E947}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3837A8AF-0D68-4552-9255-058006EE85D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3837A8AF-0D68-4552-9255-058006EE85D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3837A8AF-0D68-4552-9255-058006EE85D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3837A8AF-0D68-4552-9255-058006EE85D4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
