--- src/Tasks/Microsoft.MSBuild.LockFile.Tasks/project.json
@@ -1,11 +1,0 @@
-﻿{
-  "dependencies": {
-    "NuGet.ProjectModel": "3.4.3"
-  },
-  "frameworks": {
-    "net452": {}
-  },
-  "runtimes": {
-    "win": {}
-  }
-}
